+++ packages/core/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @grid-table/core
 
+## 0.6.5
+
+### Patch Changes
+
+- 自动调整宽度
+
 ## 0.6.4
 
 ### Patch Changes